// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io
export interface LoginData {
    email:   string;
    password: string;
}

export interface LoginResponse {
    data: User;
    userType: number;
    email: string;
}

export interface User {
    refresh: string;
    access: string;
}


export interface ProfessionalResponse {
    id:           number;
    password:     string;
    last_login:   null;
    email:        string;
    fname:        string;
    lname:        string;
    is_admin:     boolean;
    is_active:    boolean;
    is_staff:     boolean;
    is_superuser: boolean;
    userType:     number;
    curp:         string;
    phone:        string;
    document1:    string;
    document2:    string;
    schedule:     string;
    is_accepted:  boolean;
}

export interface Pro {
    fname: string;
    lname: string;
    phone: string;
    schedule: string;
}

// Generated by https://quicktype.io

export interface RegisterResponse {
    data:  RegisterData;
    token: User;
}

export interface RegisterData {
    id:           number;
    password:     string;
    last_login:   null;
    email:        string;
    fname:        string;
    lname:        string;
    is_admin:     boolean;
    is_active:    boolean;
    is_staff:     boolean;
    is_superuser: boolean;
    userType:     number;
    phone:        string;
    description:  string;
}